refactor: Decouple DI and enforce datasource packaging for localstorage
