--- core/localstorage/src/main/java/com/hilingual/core/localstorage/constant/DataStoreConstant.kt
@@ -1,23 +1,0 @@
-/*
- * Copyright 2025 The Hilingual Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hilingual.core.localstorage.constant
-
-object DataStoreConstant {
-    const val HILINGUAL_USER_INFO_PREFS = "hilingual_user_info_prefs"
-    const val ENCRYPTED_USER_PREFS = "encrypted_user_prefs.bin"
-    const val HILINGUAL_DIARY_TEMP_PREFS = "hilingual_diary_temp_prefs"
-    const val HILINGUAL_ONBOARDING_STATE_PREFS = "hilingual_onboarding_state_prefs"
-}
