--- core/localstorage/src/main/java/com/hilingual/core/localstorage/di/qualifier/OnboardingDataStore.kt
@@ -1,7 +1,0 @@
-package com.hilingual.core.localstorage.di.qualifier
-
-import javax.inject.Qualifier
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class OnboardingDataStore
