+++ data/auth/src/main/java/com/hilingual/data/auth/datasource/AuthDataStore.kt
@@ -1,0 +1,7 @@
+package com.hilingual.data.auth.datasource
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class AuthDataStore
