+++ data/diary/src/main/java/com/hilingual/data/diary/datasourceimpl/DiaryTempDataSourceImpl.kt
@@ -1,19 +1,4 @@
-/*
- * Copyright 2025 The Hilingual Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hilingual.data.diary.localstorage
+package com.hilingual.data.diary.datasourceimpl
 
 import android.net.Uri
 import androidx.datastore.core.DataStore
@@ -22,8 +7,9 @@ import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
-import com.hilingual.core.localstorage.di.qualifier.DiaryTempDataStore
-import com.hilingual.core.localstorage.util.InternalImageStorage
+import com.hilingual.data.diary.datasource.DiaryTempDataSource
+import com.hilingual.data.diary.datasourceimpl.util.InternalImageStorage
+import com.hilingual.data.diary.di.DiaryTempDataStore
 import kotlinx.coroutines.flow.first
 import java.time.LocalDate
 import javax.inject.Inject
