+++ data/diary/src/main/java/com/hilingual/data/diary/datasourceimpl/util/InternalImageStorage.kt
@@ -1,19 +1,4 @@
-/*
- * Copyright 2025 The Hilingual Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hilingual.core.localstorage.util
+package com.hilingual.data.diary.datasourceimpl.util
 
 import android.content.Context
 import android.net.Uri
