+++ data/diary/src/main/java/com/hilingual/data/diary/di/DataSourceModule.kt
@@ -16,9 +16,9 @@
 package com.hilingual.data.diary.di
 
 import com.hilingual.data.diary.datasource.DiaryRemoteDataSource
+import com.hilingual.data.diary.datasource.DiaryTempDataSource
 import com.hilingual.data.diary.datasourceimpl.DiaryRemoteDataSourceImpl
-import com.hilingual.data.diary.localstorage.DiaryTempDataSource
-import com.hilingual.data.diary.localstorage.DiaryTempDataSourceImpl
+import com.hilingual.data.diary.datasourceimpl.DiaryTempDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
