+++ data/diary/src/main/java/com/hilingual/data/diary/repository/DiaryTempRepository.kt
@@ -1,0 +1,16 @@
+package com.hilingual.data.diary.repository
+
+import android.net.Uri
+import java.time.LocalDate
+
+interface DiaryTempRepository {
+    suspend fun isDiaryTempExist(selectedDate: LocalDate): Result<Boolean>
+
+    suspend fun saveDiary(selectedDate: LocalDate, text: String, imageUri: Uri?): Result<Unit>
+
+    suspend fun getDiaryText(selectedDate: LocalDate): Result<String?>
+
+    suspend fun getDiaryImageUri(selectedDate: LocalDate): Result<String?>
+
+    suspend fun clearDiaryTemp(selectedDate: LocalDate): Result<Unit>
+}
