+++ data/onboarding/src/main/java/com/hilingual/data/onboarding/di/OnboardingDataStore.kt
@@ -1,7 +1,7 @@
-package com.hilingual.data.auth.di
+package com.hilingual.data.onboarding.di
 
 import javax.inject.Qualifier
 
 @Qualifier
 @Retention(AnnotationRetention.BINARY)
-annotation class AuthDataStore
+annotation class OnboardingDataStore
