+++ data/user/src/main/java/com/hilingual/data/user/datasourceimpl/UserLocalDataSourceImpl.kt
@@ -1,9 +1,10 @@
-package com.hilingual.data.user.localstorage
+package com.hilingual.data.user.datasourceimpl
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
+import com.hilingual.data.user.datasource.UserLocalDataSource
 import com.hilingual.core.localstorage.di.qualifier.UserInfoDataStore
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
